// WEB/matcha/FE/src/types/tag.enum.ts
export enum GenderType {
  MALE = "MALE",
  FEMALE = "FEMALE",
}

export enum PreferenceType {
  HETEROSEXUAL = "HETEROSEXUAL",
  HOMOSEXUAL = "HOMOSEXUAL",
  BISEXUAL = "BISEXUAL",
  ASEXUAL = "ASEXUAL",
}

export enum InterestType {
  SPORTS = "SPORTS",
  MUSIC = "MUSIC",
  MOVIES = "MOVIES",
  BOOKS = "BOOKS",
  GAME = "GAME",
  TRAVEL = "TRAVEL",
  FOOD = "FOOD",
  FASHION = "FASHION",
  ART = "ART",
  PHOTOGRAPHY = "PHOTOGRAPHY",
  ANIMALS = "ANIMALS",
  NATURE = "NATURE",
  SCIENCE = "SCIENCE",
  TECHNOLOGY = "TECHNOLOGY",
  POLITICS = "POLITICS",
  HEALTH = "HEALTH",
  FITNESS = "FITNESS",
  EDUCATION = "EDUCATION",
  BUSINESS = "BUSINESS",
  FINANCE = "FINANCE",
  MARKETING = "MARKETING",
  ENTREPRENEURSHIP = "ENTREPRENEURSHIP",
  SOCIAL = "SOCIAL",
  CHARITY = "CHARITY",
  VOLUNTEERING = "VOLUNTEERING",
  RELIGION = "RELIGION",
  SPIRITUALITY = "SPIRITUALITY",
  FAMILY = "FAMILY",
  FRIENDS = "FRIENDS",
  LOVE = "LOVE",
  RELATIONSHIPS = "RELATIONSHIPS",
  DATING = "DATING",
  MARRIAGE = "MARRIAGE",
  RUNNING = "RUNNING",
}

export enum ageType {
  TEENS_AND_BELOW = "TEENS_AND_BELOW",
  TWENTIES = "TWENTIES",
  THIRTIES = "THIRTIES",
  FORTIES = "FORTIES",
  FIFTIES = "FIFTIES",
  SIXTIES_AND_ABOVE = "SIXTIES_AND_ABOVE",
}

export enum RateType {
  ONE = "ONE",
  TWO = "TWO",
  THREE = "THREE",
  FOUR = "FOUR",
  FIVE = "FIVE",
}

export enum AlarmType {
  LIKED = "LIKED",
  VISITED = "VISITED",
  MESSAGED = "MESSAGED",
  MATCHED = "MATCHED",
  DISLIKED = "DISLIKED",
}

export enum SortType {
  ascAge = "ascAge",
  descAge = "descAge",
  ascRate = "ascRate",
  descRate = "descRate",
}